BLE: Move gap public header at the root.
