+++ connectivity/FEATURE_BLE/include/ble/gap/AdvertisingDataBuilder.h
@@ -27,8 +27,8 @@
 #include "ble/common/ble/UUID.h"
 #include "ble/common/ble/BLETypes.h"
 #include "ble/common/ble/blecommon.h"
-#include "ble/common/ble/gap/AdvertisingDataTypes.h"
-#include "ble/common/ble/gap/Types.h"
+#include "ble/gap/AdvertisingDataTypes.h"
+#include "ble/gap/Types.h"
 
 namespace ble {
