+++ connectivity/FEATURE_BLE/include/ble/gap/AdvertisingParameters.h
