+++ connectivity/FEATURE_BLE/include/ble/gap/Events.h
