+++ connectivity/FEATURE_BLE/include/ble/gap/Types.h
